refactor: utils 폴더 삭제하고 common으로 공통 컴포넌트 이동
