+++ frontend/src/components/common/SearchBar.tsx
